requirements.txt
hopsworks[python]
scikit-learn==1.3.2
tensorflow-cpu==2.15.0
shap
xgboost
lightgbm
python-dotenv
openmeteo-requests
requests-cache
retry-requests
tzlocal>=5.0
sqlalchemy>=2.0.0
retrying>=1.3.3
PyMySQL[rsa]>=1.0.0
pyjks>=20.0.0
mock>=5.0.0
furl>=2.1.3
boto3>=1.28.0
statsmodels>=0.14.0
pandas==2.1.4
numpy>=1.24.0
pyyaml>=6.0
joblib
fastapi>=0.104.0
pydantic>=2.0.0,<3.0.0
uvicorn
streamlit
requests
aiofiles
matplotlib
seaborn
httpx
python-multipart
plotly
confluent-kafka
